// src/data_model/enums.ts
export enum GameType{
    TEXAS_HOLD_EM = "Texas Hold'em Poker",
    OMAHA = "Omaha",
    HORSE = "Horse",
    SEVEN_CARD_STUD = "7 Card Stud",
    SEVEN_CARD_STUD_EIGHT = "7 Card Stud 8",
}

export enum LimitType {
    NO_LIMIT = "No Limit",
    FIXED_LIMIT = "Fixed Limit",
    POT_LIMIT = "Pot Limit",
    POT_LIMIT_PRE = "Pot Limit Preflop",
    MIXED_LIMIT = "Mixed Limit",
    SPREAD_LIMIT = "Spread Limit",
}